move diff error from base model to model
