+++ golem-common/src/base_model/diff/mod.rs
@@ -12,10 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub mod error;
 pub mod hash;
 
-pub use error::DiffError;
 pub use hash::Hash;
 
 pub const DIFF_MODEL_VERSION: u32 = 1;
