+++ golem-common/src/model/diff/error.rs
